Moved code from cpptest-main.h into TestMain.cpp
